// src/types/types.ts
export interface Recurso{
    anoMes: number;
    codigoPessoa: string;
    nomePessoa: string;
    tipoPessoa: string;
    municipioPessoa: string;
    siglaUFPessoa: string;
    codigoUG: string;
    nomeUG: string;
    codigoOrgao: string;
    nomeOrgao: string;
    codigoOrgaoSuperior: string;
    nomeOrgaoSuperior: string;
    valor: number;
}